Refactor user defined triton kernel tests
[ghstack-poisoned]
